map-word-embeddings/requirements.txt
PyYAML
click
progressbar2
lxml
tensorflow
scipy
XlsxWriter
backports.tempfile
sklearn